Add PHP CodeSniffer as a development dependency in composer.json
This commit adds PHP CodeSniffer to the development dependencies of the project. The `squizlabs/php_codesniffer` package is included in the `require-dev` section of the `composer.json` file to help enforce coding standards and improve code quality.

Changes include:
- Added `squizlabs/php_codesniffer` with a wildcard version constraint to the `require-dev` section of `composer.json`.

This addition will facilitate code standard checks during the development process and ensure adherence to best practices.